// lib/types.ts
export enum AvailabilityType {
  CLINIC = 'clinic',
  ONLINE = 'online',
}

export enum SlotType {
  MORNING = 'morning',
  AFTERNOON = 'afternoon',
  EVENING = 'evening',
}

export enum UserRole {
  ADMIN = 'admin',
  DOCTOR = 'doctor',
  RECEPTIONIST = 'receptionist',
  PATIENT = 'patient',
  CLINIC='clinic',
  CLINICDOCTOR='clinic_doctor',
  CLINICRECEPTIONIST='clinic_receptionist',
  SUPERADMIN='superadmin',
  INDIVIDUALDOCTOR='individual_doctor'
}